Added more opcode executors.
Almost finished with all the executors, 2 are left.
0xF00A opcode is problematic, because it needs to wait for a user input.
0xD000 opcode is related to graphics.
I will work on those opcodes while working on the machine cmd.

Added fonts sprites to the machine.

+++ pkg/disasm/disasm.go
@@ -9,7 +9,7 @@ import (
 	"github.com/eladciuraru/chip8/pkg/machine"
 )
 
-const InstructionSize uint = machine.InstructionSize
+const InstructionSize uint = uint(machine.InstructionSize)
 
 
 type Disasm struct {
